Checkings added
Checkings in the constructor method (when setting analysis structure). The program makes a copy of the default analysis structure (line82), modifies the fields which match with the user analysis structure entry (line 88), and assign the modified analysis to the vocoder object (line 91). This approach was chosen to avoid computing analysis filters each time the each field of the analysis structure were modified.